Use size_t in loops; fix DNI parsing
Replace int with size_t for patient/kinesiologist iteration to match container counts and avoid signed/unsigned issues (HijaVerPacientes.cpp, VerKinesiologosHija.cpp). Fix DNI extraction when reserving a turn by converting the Kinesiologo DNI string to integer with stoi instead of using getMatricula() (ReservarTurnoHija.cpp). Update IDE project settings (ProyectoFinalPoo.zpr) — files opened, active configuration, current source, cursor positions, and build debug/optimization levels. Binary data files for kinesiologos, pacientes and turnos were also updated.

+++ ProyectoFinalPOO/HijaVerPacientes.cpp
@@ -6,7 +6,7 @@
 
 HijaVerPacientes::HijaVerPacientes(wxWindow *parent, Consultorio *consultorio) : 
 	VerPacienteHija(parent), m_consultorio(consultorio) {
-	for(int i=0 ; i<m_consultorio->getCantPacientes() ; i++) { 
+	for(size_t i=0 ; i<m_consultorio->getCantPacientes() ; i++) { 
 		Paciente *p = m_consultorio->buscarPacientePorInd(i);
 		m_grillaPacientes->AppendRows();
 		m_grillaPacientes->SetCellValue(i,0, p->getNombre());
@@ -36,7 +36,7 @@ void HijaVerPacientes::RefrescarGrillaPacientes(){
 	if (m_grillaPacientes->GetNumberRows() !=0){
 			m_grillaPacientes->DeleteRows(0,m_grillaPacientes->GetNumberRows());
 	}
-	for(int i=0 ; i<m_consultorio->getCantPacientes() ; i++) { 
+	for(size_t i=0 ; i<m_consultorio->getCantPacientes() ; i++) { 
 		Paciente *p = m_consultorio->buscarPacientePorInd(i);
 		m_grillaPacientes->AppendRows();
 		m_grillaPacientes->SetCellValue(i,0, p->getNombre());
